refactor(stripe): move to new Fragment builder system
